feat: implement pencil tool functionality
